# Optimize findIndex, findIndexEnd and map