This makes the suggested change from https://github.com/haskell/bytestring/issues/338 so that constant arguments are not passed to helper functions.  Here are the benchmarks for findIndex and findIndexEnd:
```
old:
findIndex1/findIndices                   mean 556.2 ns  ( +- 8.501 ns  )
findIndex1/find                          mean 645.2 ns  ( +- 8.517 ns  )
findIndex1End/findIndexEnd               mean 9.832 ns  ( +- 115.2 ps  )
findIndex1End/elemIndexInd               mean 557.8 ns  ( +- 8.507 ns  )
```

```
new:
findIndex1/findIndices                   mean 526.4 ns  ( +- 12.05 ns  )
findIndex1/find                          mean 416.7 ns  ( +- 6.734 ns  )
findIndex1End/findIndexEnd               mean 9.355 ns  ( +- 111.5 ps  )
findIndex1End/elemIndexInd               mean 477.8 ns  ( +- 5.754 ns  )
```

Here is the benchmarks for map:
```
old:
traversals/map (+1)                      mean 393.1 μs  ( +- 22.82 μs  )
traversals/map (+1)                      mean 90.71 ns  ( +- 1.528 ns  )
```

```
new:
traversals/map (+1)                      mean 55.09 μs  ( +- 3.310 μs  )
traversals/map (+1)                      mean 23.46 ns  ( +- 539.6 ps  )
```