@Boarders 2021-01-15
Here is the relevant piece of core that @Bodigrim was curious about:
```haskell
Data.ByteString.$wmap [InlPrag=NOUSERINLINE[2]]
  :: (Word8 -> Word8)
     -> GHC.Prim.Addr#
     -> GHC.ForeignPtr.ForeignPtrContents
     -> GHC.Prim.Int#
     -> ByteString
[GblId,
 Arity=4,
 Str=<L,C(U(U))><L,U><L,U><L,U>,
 Unf=Unf{Src=<vanilla>, TopLvl=True, Value=True, ConLike=True,
         WorkFree=True, Expandable=True, Guidance=IF_ARGS [60 0 0 0] 171 0}]
Data.ByteString.$wmap
  = \ (w_sg45 :: Word8 -> Word8)
      (ww_sg49 :: GHC.Prim.Addr#)
      (ww1_sg4a :: GHC.ForeignPtr.ForeignPtrContents)
      (ww2_sg4b :: GHC.Prim.Int#) ->
      case GHC.Magic.runRW#
             @ ('GHC.Types.TupleRep
                  '[ 'GHC.Types.TupleRep '[], 'GHC.Types.LiftedRep])
             @ (# GHC.Prim.State# GHC.Prim.RealWorld, ByteString #)
             (\ (s_i7oX [OS=OneShot] :: GHC.Prim.State# GHC.Prim.RealWorld) ->
                case GHC.Prim.<# ww2_sg4b 0# of {
                  __DEFAULT ->
                    case GHC.Prim.newPinnedByteArray#
                           @ GHC.Prim.RealWorld ww2_sg4b s_i7oX
                    of
                    { (# ipv_i6vL, ipv1_i6vM #) ->
                    let {
                      ipv2_s9Jh :: GHC.Prim.Addr#
                      [LclId]
                      ipv2_s9Jh
                        = GHC.Prim.byteArrayContents#
                            (ipv1_i6vM
                             `cast` (UnsafeCo representational (GHC.Prim.MutableByteArray#
                                                                  GHC.Prim.RealWorld) GHC.Prim.ByteArray#
                                     :: GHC.Prim.MutableByteArray# GHC.Prim.RealWorld
                                        ~R# GHC.Prim.ByteArray#)) } in
                    join {
                      $w$j_sg3X [InlPrag=NOUSERINLINE[2], Dmd=<C(S),C(U)>]
                        :: GHC.Prim.State# GHC.Prim.RealWorld
                           -> (# GHC.Prim.State# GHC.Prim.RealWorld, ByteString #)
                      [LclId[JoinId(1)], Arity=1, Str=<L,U>]
                      $w$j_sg3X (w1_sg3V [OS=OneShot]
                                   :: GHC.Prim.State# GHC.Prim.RealWorld)
                        = let {
                            ipv3_s9Jg :: GHC.ForeignPtr.ForeignPtrContents
                            [LclId, Unf=OtherCon []]
                            ipv3_s9Jg = GHC.ForeignPtr.PlainPtr ipv1_i6vM } in
                          case GHC.Prim.touch#
                                 @ 'GHC.Types.LiftedRep
                                 @ GHC.ForeignPtr.ForeignPtrContents
                                 ipv3_s9Jg
                                 w1_sg3V
                          of s'_i6w1
                          { __DEFAULT ->
                          case GHC.Prim.touch#
                                 @ 'GHC.Types.LiftedRep
                                 @ GHC.ForeignPtr.ForeignPtrContents
                                 ww1_sg4a
                                 s'_i6w1
                          of s'1_i7pa
                          { __DEFAULT ->
                          (# s'1_i7pa,
                             Data.ByteString.Internal.BS ipv2_s9Jh ipv3_s9Jg ww2_sg4b #)
                          }
                          } } in
                    joinrec {
                      $wmap__sg44 [InlPrag=NOUSERINLINE[2], Occ=LoopBreaker]
                        :: GHC.Prim.Int#
                           -> GHC.Prim.State# GHC.Prim.RealWorld
                           -> (# GHC.Prim.State# GHC.Prim.RealWorld, ByteString #)
                      [LclId[JoinId(2)], Arity=2, Str=<L,U><L,U>, Unf=OtherCon []]
                      $wmap__sg44 (ww3_sg42 :: GHC.Prim.Int#)
                                  (w1_sg3Z :: GHC.Prim.State# GHC.Prim.RealWorld)
                        = case GHC.Prim.>=# ww3_sg42 ww2_sg4b of {
                            __DEFAULT ->
                              case GHC.Prim.readWord8OffAddr#
                                     @ GHC.Prim.RealWorld
                                     (GHC.Prim.plusAddr# ww_sg49 ww3_sg42)
                                     0#
                                     w1_sg3Z
                              of
                              { (# ipv4_i7qw, ipv5_i7qx #) ->
                              case w_sg45 (GHC.Word.W8# ipv5_i7qx) of { W8# x_i7tG ->
                              case GHC.Prim.writeWord8OffAddr#
                                     @ GHC.Prim.RealWorld
                                     (GHC.Prim.plusAddr# ipv2_s9Jh ww3_sg42)
                                     0#
                                     x_i7tG
                                     ipv4_i7qw
                              of s2_i7tI
                              { __DEFAULT ->
                              jump $wmap__sg44 (GHC.Prim.+# ww3_sg42 1#) s2_i7tI
                              }
                              }
                              };
                            1# -> jump $w$j_sg3X w1_sg3Z
                          }; } in
                    jump $wmap__sg44 0# ipv_i6vL
                    };
                  1# ->
                    case GHC.ForeignPtr.mallocPlainForeignPtrBytes2 of wild_00 { }
                })
      of
      { (# ipv_i6Ap, ipv1_i6Aq #) ->
      ipv1_i6Aq
      }
```

It does indeed have the join points that you had mentioned wanting to see.