@Boarders 2021-01-15
Ah, I realised that the accidentally missing INLINE was making a big difference and leading to things that are too good to be true. Once I add in INLINE here is the performance I get from the old and new map:
```
old:
traversals/map (+1)                      mean 374.0 μs  ( +- 8.434 μs  )
traversals/map (+1)                      mean 88.36 ns  ( +- 2.557 ns  )
```

```
new:
traversals/map (+1)                      mean 254.8 μs  ( +- 21.48 μs  )
traversals/map (+1)                      mean 62.21 ns  ( +- 2.377 ns  )
```
If I remove the INLINE pragma from each definition then I get:
```
old:
traversals/map (+1)                      mean 37.68 μs  ( +- 2.459 μs  )
traversals/map (+1)                      mean 18.71 ns  ( +- 263.0 ps  )
```

```
new:
traversals/map (+1)                      mean 36.49 μs  ( +- 910.0 ns  )
traversals/map (+1)                      mean 19.23 ns  ( +- 420.8 ps  )
```

This is presumably because when we don't have INLINE then both definitions can correctly get inlined even if unsaturated. This leads me to think that we could really, really do with having a look at using INLINABLE over INLINE on some of these functions as I think writing:

```haskell
map f $ bs
```
over 
```haskell
map f bs
```
should not lead to a potentially 6x performance degradation.